+++ src/stores/anime-store.ts
@@ -1,7 +1,7 @@
 import { defineStore } from 'pinia';
 import { Anime } from 'src/components/models';
 
-export const useAnimeStore = defineStore('animeStore', {
+export const useAnimeStore = defineStore('anime', {
     state: () => ({
         anime: {} as Anime,
     }),
