+++ src/stores/poster-anime-store.ts
@@ -1,6 +1,6 @@
 import { defineStore } from 'pinia';
 
-export const usePosterAnimeStore = defineStore('posterAnimeStore', {
+export const usePosterAnimeStore = defineStore('poster-anime', {
     state: () => ({
         anime: {
             seasonal: [],
